chore(): update tsconfig path settings
Many tools rely on tsconfig in different ways, some supporting the
'files' and 'exclude' properties differently, as well as the nonstandard
'filesGlob'.  This is an attempt to have awesome-typescript-loader
building as quickly as possible, while removing IDE errors about files
not being in the compilation context.

Note: as of 0.15.10 there appear to be some bugs/inconsistencies in the
way awesome-typescript-loader resolves paths between files, exclude
and filesGlob when the resolveGlobs option set to false vs true.

Related: https://github.com/driftyco/ionic-cli/issues/741.